Fix the type of product stock

+++ WebAPI/Models/Product.cs
@@ -7,7 +7,7 @@
         public string? Description { get; set; }
         public decimal Price { get; set; }
         public string? UrlImage { get; set; }
-        public int Stock { get; set; }
+        public float Stock { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime UpdatedAt { get; set; }
         public int CategoryId { get; set; }
